More work on converting game to use Fleet.
A fleet properly moves and updates itself now.
The Status menu supports viewing multiple ships!

+++ ChartedWaters/State_Combat.cpp
@@ -65,7 +65,7 @@ void CombatMap::gen()
         temp += 0.7;
       temp += 0.0;
       // buffer[xcounter + ycounter * w] = temp > 0 ? temp * 20 : 0;
-      ref(xcounter, ycounter).moisture = temp;
+      ref(xcounter, ycounter).moisture = (float)temp;
       }
 
     setCoastFlags();
